Move png functions
